major rework of test setup (work in progress)

+++ util/bin/testSetup.sh
@@ -29,9 +29,9 @@ function buildDistServer {
 #
 # DANGER: remove installation without (almost) any trace
 function cleanup {
-    node=`echo $1 | cut -d' ' -f1`
-    remote=`echo $1 | cut -d' ' -f2`
-    login=`echo $1 | cut -d' ' -f3`
+    node=`echo $1 | cut -d';' -f1`
+    remote=`echo $1 | cut -d';' -f2`
+    login=`echo $1 | cut -d';' -f3`
     echo "performing teardown at $login@remote ($node) ..."
     ssh -o "StrictHostKeyChecking no" $login@$remote "./configBatch.sh CLEANUP"
 }
@@ -40,23 +40,24 @@ function cleanup {
 #
 #
 function compile {
-
-    mvn --batch-mode -DskipTests clean install
-
-    pushd $LBAC_REPO/ui >/dev/null
-    REVISION=`mvn org.apache.maven.plugins:maven-help-plugin:evaluate -Dexpression=project.version -q -DforceStdout`
-    MAJOR=`echo $REVISION | cut -d. -f1`
-    MINOR=`echo $REVISION | cut -d. -f2`
-    RELEASE="$MAJOR.$MINOR"
-    popd >/dev/null
+    stage=$1
+    echo "=== Build Docker Images ==="
+    if [ -n $stage ] ; then
+        $LBAC_REPO/util/bin/buildDocker.sh \
+            --branch-file $BRANCH_FILE \
+            --stage-label $stage \
+            --test-registry `hostname -f`:5000
+    else 
+        $LBAC_REPO/util/bin/buildDocker.sh --test-registry `hostname -f`:5000
+    fi
 }
 
 #
 #
 #
 function copyNodeConfig {
-    cloud=`echo $1 | cut -d' ' -f1`
-    node=`echo $1 | cut -d' ' -f2`
+    cloud=`echo $1 | cut -d';' -f1`
+    node=`echo $1 | cut -d';' -f2`
     cp $LBAC_REPO/config/nodes/${node}_${cloud}.sh.asc $LBAC_REPO/config/$cloud/
 }
 
@@ -64,21 +65,26 @@ function copyNodeConfig {
 #
 #
 function createNodeConfig {
-    node=`echo $1 | cut -d' ' -f1`
-    remote=`echo $1 | cut -d' ' -f2`
-    login=`echo $1 | cut -d' ' -f3`
-    cloud=`grep $node "$LBAC_REPO/util/test/etc/nodeconfig.txt" | \
-        cut -c9-18 | \
-        sed -e 's/^[[:blank:]]*//;s/[[:blank:]]*$//'`
+    node=`echo $1 | cut -d';' -f1`
+    remote=`echo $1 | cut -d';' -f2`
+    login=`echo $1 | cut -d';' -f3`
+    cloud=`grep $node "$LBAC_REPO/util/test/etc/nodeconfig.cfg" | \
+        grep PRIMARY_CLOUD | \
+        cut -d= -f2`
     crimsyhost=`hostname -f`
 
     echo "executing createNodeConfig for host $login@$remote ($node) ..."
+
+    echo "copying node configuration ..."
+    grep -E "^$node;" "$LBAC_REPO/util/test/etc/nodeconfig.cfg" | \
+    cut -d';' -f2- |\
+    ssh -o "StrictHostKeyChecking no" $login@$remote bash -c "cat - > nodeconfig.cfg"
 
     echo "copying configBatch.sh script ..."
     scp -q -o "StrictHostKeyChecking no" "$LBAC_REPO/util/bin/configBatch.sh" $login@$remote:
 
     echo "executing configBatch.sh ..."
-    ssh -o "StrictHostKeyChecking no" $login@$remote "chmod +x configBatch.sh && ./configBatch.sh CONFIG $crimsyhost $cloud $node"
+    ssh -o "StrictHostKeyChecking no" $login@$remote "chmod +x configBatch.sh && ./configBatch.sh CONFIG $crimsyhost"
 
     echo "fetching node configuration ..."
     scp -q -o "StrictHostKeyChecking no" $login@$remote:etc/$cloud/config.sh.asc "$LBAC_REPO/config/nodes/${node}_${cloud}.sh.asc"
@@ -110,9 +116,36 @@ function error {
 #
 #
 #
+function infoLog {
+    msg=`echo "$1 FINISHED                                                        -" | cut -c-63`
+    echo "*****************************************************************"
+    echo "*                                                               *"
+    echo "* $msg *"
+    echo "* Please find the log file in config/logs/test.$TEST_DATE.log *"
+    echo "*                                                               *"
+    echo "*****************************************************************"
+}
+
+#
+#
+#
+function installFunc {
+    # install node
+    echo "=== install nodes ==="
+    cat $HOSTLIST |\
+    if [$NODE = "all" ] ; then cat ; else grep $NODE ; fi |\
+     while read record ; do
+        echo | installNode "$record"
+    done
+
+}
+
+#
+#
+#
 function installNode {
-    remote=`echo $1 | cut -d' ' -f2`
-    login=`echo $1 | cut -d' ' -f3`
+    remote=`echo $1 | cut -d';' -f2`
+    login=`echo $1 | cut -d';' -f3`
     echo "installNode called for host: $login@$remote"
     ssh -o "StrictHostKeyChecking no" $login@$remote "./bin/install.sh"
 }
@@ -125,16 +158,16 @@ function restore {
         echo "Restoring snapshot '$RESTORE' on all nodes"
         cat $HOSTLIST | \
             while read record ; do
-                remote=`echo $record | cut -d' ' -f2`
-                login=`echo $record | cut -d' ' -f3`
+                remote=`echo $record | cut -d';' -f2`
+                login=`echo $record | cut -d';' -f3`
                 ssh -o "StrictHostKeyChecking no" "$login@$remote" ./dist/bin/setupROOT.sh restore $RESTORE
             done
     else 
         echo "Restoring snapshot '$RESTORE' on node '$RESTORE'"
         grep $NODE $HOSTLIST | \
             while read record ; do
-                remote=`echo $record | cut -d' ' -f2`
-                login=`echo $record | cut -d' ' -f3`
+                remote=`echo $record | cut -d';' -f2`
+                login=`echo $record | cut -d';' -f3`
                 ssh -o "StrictHostKeyChecking no" "$login@$remote" ./dist/bin/setupROOT.sh restore $RESTORE
             done
     fi
@@ -144,7 +177,10 @@ function restore {
 #
 #
 function runDistServer {
-    echo "Setting up distribution servers"
+    echo "=== Distribution Servers ===" 
+    buildDistServer
+
+    mkdir -p "$LBAC_REPO/target/integration/htdocs/$dist"
     cp docker/crimsyci/index.html target/integration/htdocs
     (docker inspect crimsyci_service | grep Status | grep -q running ) && docker stop crimsyci_service
     docker inspect crimsyci_service >/dev/null 2>&1 && docker rm crimsyci_service 
@@ -163,22 +199,73 @@ function runDistServer {
 }
 
 #
+#
+#
+function runSetup {
+    runDistServer
+    if [ -n $BRANCH_FILE ] ; then
+        initial_stage=`cut -d';' -f1 $BRANCH_FILE | sort | uniq | head -1`
+        echo "Executing initial setup stage $initial_stage"
+    else 
+        echo "Executing direct setup"
+        initial_stage=''
+    fi
+    compile $initial_stage
+    setupFunc
+    installFunc
+
+    # ToDo: multiple cloud memberships 
+
+    echo "sleep 15 seconds to settle everything ..."
+    sleep 15
+
+    if [ -n $BRANCH_FILE ] ; then
+        cut -d';' -f1 $BRANCH_FILE |\
+        grep -v $initial_stage | sort | uniq |\
+        while read record ; do
+            echo "compiling for stage $record"
+            compile $record
+            UPDATE_CMD=""
+            runUpdate 
+            echo "sleep 15 seconds to settle everything ..."
+            sleep 15
+        done
+    fi
+}
+
+#
 # run Tests. 
-# NOTE: currently only runs on first node in HOSTLIST
+# NOTE: tests are currently available for node1 only 
 #
 function runTests {
     # check prerequisites
-    echo "checking prerequsites"
-    (docker inspect dist_proxy_1 2>/dev/null | grep -q running ) \
-        || error "Service seems unavailable ..."
+    echo "checking test prerequsites ..."
+    cat $HOSTLIST |\
+    if [ $NODE = "all" ] ; then cat ; else grep $NODE ; fi |\
+    while read record ; do
+        node=`echo $record | cut -d';' -f1`
+        remote=`echo $record | cut -d';' -f2`
+        login=`echo $record | cut -d';' -f3`
+        echo | ssh -o "StrictHostKeyChecking no" "$login@$remote" \
+            bash -c "docker inspect dist_proxy_1 2>/dev/null | grep -q running" \
+            || error "Service seems unavailable at node $node"
+
+    done
 
     # initialize database with test data
-    HOST=`head -1 $HOSTLIST | cut -d' ' -f2`
-    docker cp $LBAC_REPO/util/test/etc/initial_data.sql dist_db_1:/tmp/
-    wget -o /dev/null -O /dev/null --no-check-certificate https://$HOST/ui/index.xhtml
-    echo "waiting 3 sec. for webapp to initialize database ..."
-    sleep 3
-    docker exec -i -u postgres dist_db_1 psql -Ulbac lbac -f /tmp/initial_data.sql
+    echo "loading initial data ..."
+    cat $HOSTLIST |\
+    if [ $NODE = "all" ] ; then cat ; else grep $NODE ; fi |\
+    while read record ; do
+        node=`echo $record | cut -d';' -f1`
+        remote=`echo $record | cut -d';' -f2`
+        login=`echo $record | cut -d';' -f3`
+        if [ -s $LBAC_REPO/util/test/etc/$node.initial_data.sql ] ; then
+            echo "processing node $node"
+            echo | scp -o "StrictHostKeyChecking no" $LBAC_REPO/util/test/etc/$node.initial_data.sql "$login@$remote:tmp/initial_data.sql"
+            echo | ssh -o "StrictHostKeyChecking no" "$login@$remote" ./configBatch LOAD_DATA
+        fi
+    done
 
     # build test containers and set up environment
     echo "checking / building test environment"
@@ -192,26 +279,38 @@ function runTests {
 dnl
 dnl Cypress test fixtures configuration
 dnl
-define(\`TESTBASE_HOSTNAME',\`$HOST')dnl
+define(\`TESTBASE_HOSTNAME',\`$remote')dnl
 EOF
     find $LBAC_REPO/target/cypress -type f -name "*.m4" -exec /bin/bash -c cypressPreprocess {} \;
 
     # run tests ...
-    echo "running tests"
+    echo "*"
+    echo "* Running Tests"
+    echo "* Test outcomes will be saved in target/cypress/"
+    echo "*"
     docker run -v $LBAC_REPO/target/cypress/cypress:/app/cypress --name cy1 cypress --browser firefox --headless
 
     # clean up
     echo "removing test container"
     docker rm cy1
     popd >/dev/null
+}
 
-    echo "*****************************************************************"
-    echo "*                                                               *"
-    echo "* TESTS FINISHED                                                *"
-    echo "* Please find the log file in config/logs/test.$TEST_DATE.log *"
-    echo "* and test outcomes in directory target/cypress/                *"
-    echo "*                                                               *"
-    echo "*****************************************************************"
+#
+#
+#
+function runUpdate {
+    echo "=== run update on nodes ==="
+    cat $HOSTLIST |\
+    if [ $NODE = "all" ] ; then cat ; else grep $NODE ; fi |\
+     while read record ; do
+        node=`echo $record | cut -d';' -f1`
+        remote=`echo $record | cut -d';' -f2`
+        login=`echo $record | cut -d';' -f3`
+
+        echo "updating node $node with $UPDATE_CMD"
+        echo | ssh -o "StrictHostKeyChecking no" "$login@$remote" ./dist/bin/update.sh $UPDATE_CMD
+    done
 }
 
 #
@@ -260,13 +359,6 @@ function setupFunc {
             setupTestCAconf "$record"
         done
 
-    echo "=== Distribution Server ==="
-    buildDistServer
-    runDistServer
-
-    echo "=== Build Docker Images ==="
-    $LBAC_REPO/util/bin/buildDocker.sh `hostname -f`:5000
-
     echo "=== Setup ROOT CA ==="
     setupTestRootCA
 
@@ -278,14 +370,16 @@ function setupFunc {
         done
 
     echo "=== create node configurations ==="
-    cat $HOSTLIST | while read record ; do
+    cat $HOSTLIST |\
+    if [ $NODE = "all" ] ; then cat ; else grep $NODE ; fi |\
+     while read record ; do
         echo | createNodeConfig "$record"
     done
 
     # package master nodes
     echo "=== determine master nodes ==="
     tail -n +2 $LBAC_REPO/util/test/etc/cloudconfig.txt | \
-        cut -d: -f1 | \
+        cut -d';' -f1 | \
         while read record ; do
             $LBAC_REPO/util/bin/package.sh "$record" MASTERBATCH
         done
@@ -297,31 +391,10 @@ function setupFunc {
             copyNodeConfig "$record"
         done
     tail -n +2 $LBAC_REPO/util/test/etc/cloudconfig.txt | \
-        cut -d: -f1 | \
+        cut -d';' -f1 | \
         while read record ; do
             $LBAC_REPO/util/bin/package.sh "$record" AUTOBATCH
         done
-
-    # install node
-    echo "=== install nodes ==="
-    cat $HOSTLIST | while read record ; do
-        echo | installNode "$record"
-    done
-
-    #
-    # ToDo: multiple cloud memberships 
-    #
-
-    #
-    echo "sleep 15 seconds to settle everything ..."
-    sleep 15
-
-    echo "*****************************************************************"
-    echo "*                                                               *"
-    echo "* SETUP FINISHED                                                *"
-    echo "* Please find the log file in config/logs/test.$TEST_DATE.log *"
-    echo "*                                                               *"
-    echo "*****************************************************************"
 }
 
 #
@@ -336,7 +409,7 @@ function setupTestRootCA {
 #
 #
 function setupTestSubCA {
-    cloud=`echo $1 | cut -d: -f1`
+    cloud=`echo $1 | cut -d';' -f1`
 
     $LBAC_REPO/util/bin/camgr.sh --batch --mode ca --cloud $cloud
 
@@ -355,12 +428,14 @@ function setupTestSubCA {
 }
 
 function setupConfigure {
-    cloud=`echo $1 | cut -d: -f1`
+    cloud=`echo $1 | cut -d';' -f1`
     LBAC_CA_DIR=$LBAC_REPO/config/$cloud/CA
     . $LBAC_CA_DIR/cloud.cfg
 
+    revision=`grep "CURRENT" $LBAC_REPO/config/revision_info.txt |\
+        cut -d';' -f1`
     sed -e "s,CLOUDCONFIG_DOWNLOAD_URL,$DOWNLOAD_URL," $LBAC_REPO/util/bin/configure.sh | \
-    sed -e "s,CLOUDCONFIG_CURRENT_RELEASE,$RELEASE," |\
+    sed -e "s,CLOUDCONFIG_CURRENT_RELEASE,$revision," |\
     sed -e "s,CLOUDCONFIG_CLOUD_NAME,$CLOUD," |\
     openssl smime -sign -signer $LBAC_CA_DIR/$DEV_CERT.pem \
       -md sha256 -binary -out $LBAC_REPO/target/integration/htdocs/$cloud/configure.sh.sig \
@@ -375,14 +450,12 @@ function setupConfigure {
 #
 #
 function setupTestCAconf {
-    dist=`echo $1 | cut -d: -f1`
-    superior=`echo $1 | cut -d: -f2`
-    cloud=`echo $1 | cut -d: -f3`
-    name=`echo $1 | cut -d: -f4`
+    dist=`echo $1 | cut -d';' -f1`
+    superior=`echo $1 | cut -d';' -f2`
+    cloud=`echo $1 | cut -d';' -f3`
+    name=`echo $1 | cut -d';' -f4`
 
     mkdir -p "$LBAC_REPO/config/$cloud/CA"
-    mkdir -p "$LBAC_REPO/target/integration/htdocs/$dist"
-    cp "$LBAC_REPO/util/test/etc/nodeconfig.txt" "$LBAC_REPO/target/integration/htdocs/$dist"
 
     cat > "$LBAC_REPO/config/$cloud/CA/cloud.cfg" <<EOF
 #
@@ -422,16 +495,16 @@ function snapshot {
         echo "Taking snapshot '$SNAPSHOT' of all nodes"
         cat $HOSTLIST | \
             while read record ; do
-                remote=`echo $record | cut -d' ' -f2`
-                login=`echo $record | cut -d' ' -f3`
+                remote=`echo $record | cut -d';' -f2`
+                login=`echo $record | cut -d';' -f3`
                 ssh -o "StrictHostKeyChecking no" "$login@$remote" ./dist/bin/setupROOT.sh snapshot $SNAPSHOT
             done
     else
         echo "Taking snapshot '$SNAPSHOT' for node '$NODE'"
         grep $NODE $HOSTLIST | \
             while read record ; do
-                remote=`echo $record | cut -d' ' -f2`
-                login=`echo $record | cut -d' ' -f3`
+                remote=`echo $record | cut -d';' -f2`
+                login=`echo $record | cut -d';' -f3`
                 ssh -o "StrictHostKeyChecking no" "$login@$remote" ./dist/bin/setupROOT.sh snapshot $SNAPSHOT
             done
     fi
@@ -443,16 +516,24 @@ function snapshot {
 function teardown {
 
     # tear down nodes
-    cat $HOSTLIST | while read record ; do
+    cat $HOSTLIST |\
+    if [ $NODE = all ] ; then cat ; else grep $NODE ; fi |\
+     while read record ; do
         echo | cleanup "$record"
     done
 
-    # remove target directory (may need root privilege?)
-    rm -rf config/ target/
+    if [ $NODE = all ] ; then
+        # remove target directory (may need root privilege?)
+        rm -rf config/ target/
 
-    docker stop crimsyci_service
-    docker stop crimsyreg_service
-    docker container prune -f
+        docker stop crimsyci_service
+        docker stop crimsyreg_service
+        docker container prune -f
+    else
+        echo 
+        echo "NOTE: teardown limited to node $NODE"
+        echo
+    fi
 
     echo "*****************************************************************"
     echo "*                                                               *"
@@ -473,8 +554,8 @@ ${BOLD}NAME${REGULAR}
 
 ${BOLD}SYNOPSIS${REGULAR}
     testSetup.sh [-H|hostlist HOSTLIST] [-h|--help] [-n|--node NODE] 
-        [-p|--pause] [-R|--restore LABEL] [-r|--runTests] [-S|--snapshot LABEL] 
-        [-s|--setup] [-t|--teardown] [-w|--wake]
+        [-p|--pause NODE] [-R|--restore LABEL] [-r|--runTests] [-S|--snapshot LABEL] 
+        [-s|--setup] [-t|--teardown] [-u|--update CMD] [-w|--wake NODE]
 
 ${BOLD}DESCRIPTION${REGULAR}
     Set up, operate and clean up a test environment of CRIMSy. Setup includes 
@@ -483,10 +564,15 @@ ${BOLD}DESCRIPTION${REGULAR}
     involved hosts.
 
 ${BOLD}OPTIONS${REGULAR}
+-b|--branch-file FILE
+    a list of branches to compile and build on each stage. Contains a stage 
+    label, the the branch name and respective flags for container image 
+    tagging separated by semicolon
+
 -H|--hostlist HOSTLIST
-    mapping of node names (node1, node2, ...) to real host names. Each line contains
-    the node name followed by the host name and the user account separated by a single 
-    space character.
+    mapping of node names (node1, node2, ...) to real host names. Each line 
+    contains the node name followed by the host name and the user account 
+    separated by semicolon.
 
 -h|--help
     print this help text
@@ -494,7 +580,7 @@ ${BOLD}OPTIONS${REGULAR}
 -n|--node NODE
     operate on node NODE only (default all)
 
--p|--pause
+-p|--pause NODE
     pause execution of node NODE
 
 -R|--restore LABEL
@@ -512,7 +598,10 @@ ${BOLD}OPTIONS${REGULAR}
 -t|--teardown
     remove all nodes and clean up everything
 
--w|--wake
+-u|--update CMD
+    run update.sh with argument
+
+-w|--wake NODE
     resume execution of node NODE
 EOF
 
@@ -535,28 +624,40 @@ function mainFunc {
 
     if [ -n "$SNAPSHOT" ] ; then
         snapshot
+        infoLog SNAPSHOT
         exit 0
     fi
 
     if [ -n "$RESTORE" ] ; then
         restore
+        infoLog RESTORE
         exit 0
     fi
 
     if [ -n "$SETUP" ] ; then
-        echo "setup"
-        compile
-        setupFunc
+        echo "Starting setup ..."
+        runSetup
+        infoLog SETUP
+    fi
+
+    if [ -n "$UPDATE" ] ; then
+        if [ -n "$SETUP" ] ; then
+            error "Cannot combine setup and update in a single run"
+        fi
+        echo "Run update ..."
+        runUpdate 
+        infoLog UPDATE
     fi
 
     if [ -n "$RUNTESTS" ] ; then
         echo "run tests"
         runTests
+        infoLog TESTS
     fi
 
     if [ -n "$TEARDOWN" ] ; then
         echo "tear down"
-        teardown
+        teardown 
     fi
 }
 #
@@ -569,6 +670,7 @@ cd $LBAC_REPO
 safetyCheck
 TEST_DATE=`date +%Y%m%d%H%M`
 
+BRANCH_FILE=''
 HOSTLIST=''
 NODE=all
 PAUSE=''
@@ -577,9 +679,11 @@ RUNTESTS=''
 SNAPSHOT=''
 SETUP=''
 TEARDOWN=''
+UPDATE=''
+UPDATE_CMD=''
 WAKE=''
 
-GETOPT=$(getopt -o 'H:hn:pR:rS:stw' --longoptions 'hostlist:,help,node:,pause,restore:,runTests:,snapshot:,setup,teardown:,wake' -n 'testSetup.sh' -- "$@")
+GETOPT=$(getopt -o 'b:H:hn:p:R:rS:stu:w:' --longoptions 'branch-file:hostlist:,help,node:,pause:,restore:,runTests,snapshot:,setup,teardown,update:wake:' -n 'testSetup.sh' -- "$@")
 
 if [ $? -ne 0 ]; then
         echo 'Error in commandline evaluation. Terminating...' >&2
@@ -591,6 +695,11 @@ unset GETOPT
 
 while true ; do
     case "$1" in
+    '-b'|'--branch-file')
+        BRANCH_FILE="$2"
+        shift 2
+        continue
+        ;;
     '-H'|'--hostlist')
         HOSTLIST=`realpath "$2"`
         shift 2
@@ -633,6 +742,12 @@ while true ; do
     '-t'|'--teardown')
         TEARDOWN='teardown'
         shift
+        continue
+        ;;
+    '-u'|'--update')
+        UPDATE='update'
+        UPDATE_CMD=$2
+        shift 2
         continue
         ;;
     '-w'|'--wake')
